Major Update to the Code
Hey guys, here’s what I did so far:

1) Cleaned up threadtest.cc so we can initialize threads and appClerks
for now the proper way
2) Fixed some errors in the synch.cc with lock code
3) Fixed a lot of the application clerk and customer code. It’s about
90% working, getting some weird lock misuse errors when waiting during
the two thread communication process. I’m thinking maybe it’s the lock
and condition implementation so if someone can look over that that
would be great. I’ll try to fix that tonight and have the appClerk and
picClerk done.