Part 3 Started
To test our new C code for the simulation, I added in all the code we
would need as comments to make it working on our old files. So pretty
much the entire code for getting the passport office to work in C is
there. All that is left to do is test and to make sure it works, which
we will do in threadtest.cc then transfer over to the userprog
directory to save us the time of debugging and what not since we
already know our synchronization tools work fine in kernel mode. So
summary:

-Added in all the C code as needed in the form of comments (except
print and input statements) to make the simulation work in C
-The new C code needs to be slowly integrated and tested to make sure
it all works
-If this new C code is tested and works as expected, all we have to
make sure is done is sys calls.