All Syscalls Implemented (Not Tested)
I have implemented all the syscalls on both the server end and client
end. I have not compiled the code though since my aludra is giving me
weird issues with syncing the code, so you guys will have to compile to
see if it works though I have ran through the logic myself and checked
as much as I code and all seems to be fine and clear. Some quick notes:
-Syscalls that return SUCCESS by the server will return -2 value to the
user program, and will return -1 value as an ERROR to the user program.
So keep this in mind when creating tests.