Fixed Some Things
-Added in the name checking for locks, CVs, and MVs
-Did not add in thread specific messaging yet since I still have a few
questions before I can implement, and also since this project may not
even require it, so wanted to make sure. As of now, the only way that I
can think we can implement would be to have a global int counter to
give an ID to each thread, so when a reply is sent to say two threads
waiting on the same message, you can specify which message is for whom
by the global id that is also sent back. Only problem that I see with
this is that if a reply is received by the wrong thread, then the reply
itself will be lost forever. Still, need to think this through more.